Primer fragmento listo

+++ app/src/main/java/cl/jpvs/mod6ej02/data/local/Item.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "table_item")
-data class Item(val nombre:String, val precio: Int, val cantidad: Int){
+data class Item(val nombre:String, val precio: Double, val cantidad: Double){
     @PrimaryKey (autoGenerate = true) var id : Long = 0
 }
